Move InternalAuthenticationManager into identity module
