+++ sandbox/libs/authn/build.gradle
@@ -14,13 +14,6 @@ apply plugin: 'opensearch.publish'
 
 
 dependencies {
-  api 'org.apache.shiro:shiro-core:1.9.1'
-  // Needed for shiro
-  implementation "org.slf4j:slf4j-api:${versions.slf4j}"
-
-  // logging
-  implementation "org.apache.logging.log4j:log4j-api:${versions.log4j}"
-
   testImplementation(project(":test:framework")) {
     exclude group: 'org.opensearch.sandbox', module: 'opensearch-authn'
   }
@@ -29,28 +22,6 @@ dependencies {
 tasks.named('forbiddenApisMain').configure {
   replaceSignatureFiles 'jdk-signatures'
 }
-
-tasks.named("dependencyLicenses").configure {
-  mapping from: /log4j-.*/, to: 'log4j'
-}
-
-thirdPartyAudit.ignoreMissingClasses(
-  'org.apache.commons.beanutils.BeanUtilsBean',
-  'org.apache.commons.beanutils.ConvertUtilsBean',
-  'org.apache.commons.beanutils.PropertyUtilsBean',
-  'org.apache.commons.beanutils.SuppressPropertiesBeanIntrospector',
-  'org.apache.commons.configuration2.interpol.ConfigurationInterpolator',
-  'org.osgi.framework.Bundle',
-  'org.osgi.framework.BundleActivator',
-  'org.osgi.framework.BundleContext',
-  'org.osgi.framework.BundleEvent',
-  'org.osgi.framework.SynchronousBundleListener',
-  'org.osgi.framework.wiring.BundleWire',
-  'org.osgi.framework.wiring.BundleWiring',
-  'org.slf4j.impl.StaticLoggerBinder',
-  'org.slf4j.impl.StaticMDCBinder',
-  'org.slf4j.impl.StaticMarkerBinder',
-)
 
 tasks.register("integTest", Test) {
   include '**/*IT.class'
