--- sandbox/libs/authn/src/main/java/org/opensearch/authn/internal/package-info.java
@@ -1,7 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-/** Classes for the internal authentication in OpenSearch */
-package org.opensearch.authn.internal;
