+++ sandbox/modules/identity/src/main/java/org/opensearch/identity/authmanager/internal/InternalAccessTokenManager.java
@@ -1,8 +1,11 @@
 /*
- * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
  */
-package org.opensearch.authn.internal;
+package org.opensearch.identity.authmanager.internal;
 
 import org.opensearch.authn.AccessTokenManager;
 import org.opensearch.authn.tokens.AccessToken;
