+++ sandbox/modules/identity/src/main/java/org/opensearch/identity/authmanager/internal/InternalSubject.java
@@ -1,14 +1,16 @@
 /*
- * Copyright OpenSearch Contributors
  * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
  */
 
-package org.opensearch.authn.internal;
+package org.opensearch.identity.authmanager.internal;
 
 import java.security.Principal;
 import java.util.Objects;
 
-import org.opensearch.authn.AuthenticationTokenHandler;
 import org.opensearch.authn.tokens.AuthenticationToken;
 import org.opensearch.authn.Subject;
