+++ sandbox/modules/identity/src/main/java/org/opensearch/identity/authmanager/internal/package-info.java
@@ -1,0 +1,10 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+/** Classes for the internal authentication in OpenSearch */
+package org.opensearch.identity.authmanager.internal;
