+++ sandbox/modules/identity/src/main/java/org/opensearch/identity/authmanager/securityplugin/SecurityPluginAuthenticationManager.java
@@ -6,7 +6,7 @@
  * compatible open source license.
  */
 
-package org.opensearch.identity.securityplugin;
+package org.opensearch.identity.authmanager.securityplugin;
 
 import org.opensearch.authn.AccessTokenManager;
 import org.opensearch.authn.AuthenticationManager;
